Improve metadata bounding box projection, move interpolateLinear to GeometyHelper
